+++ current/bloom/KCountArray6MT.java
@@ -532,7 +532,7 @@ public class KCountArray6MT extends KCountArray {
 			int cellShift=(int)(cellBits*key);
 			int value=((word>>>cellShift)&valueMask);
 			if(value==0){cellsUsedPersonal++;}
-			value=min(value+1, maxValue);
+			value=(int)min(value+1L, maxValue);
 			word=(value<<cellShift)|(word&~((valueMask)<<cellShift));
 			array[index]=word;
 			return value;
