+++ current/bloom/KCountArray7MT.java
@@ -617,7 +617,7 @@ public class KCountArray7MT extends KCountArray {
 			int cellShift=(int)(cellBits*key);
 			int value=((word>>>cellShift)&valueMask);
 			if(value==0){cellsUsedPersonal++;}
-			value=min(value+1, maxValue);
+			value=(int)min(value+1L, maxValue);
 			word=(value<<cellShift)|(word&~((valueMask)<<cellShift));
 			array[index]=word;
 			return value;
